requirements.txt
requests>=2.28.2
urllib3>=1.26.14
python-dotenv>=0.21.1
dnacentersdk>=2.5.6
textfsm~=1.1.2
netmiko~=4.1.2
pyats~=23.1
--extra-index-url https://artifactory.nike.com/artifactory/api/pypi/python-virtual/simple
netdevops~=1.0.100
mydict~=2.1.0
openpyxl~=3.1.1
reportlab~=3.6.12
PyPDF2~=3.0.1
maclookup~=1.0.3
proxmoxer~=2.0.1
python-dotenv~=0.21.1
pyats[full]
